refactor: migrate text strings to AppStrings for localization
- Replaced instances of DesignSystem.Text with AppStrings in various views and view models to facilitate localization.
- Updated MainTabView, QuizViewModel, ScoreView, QuizTabView, ResultView, and WelcomeView to use the new AppStrings structure for better maintainability and internationalization support.
